refactor(grid): export props type
