+++ packages/grid/docs/grid.mdx
@@ -21,7 +21,7 @@ import * as packageJson from '../package.json';
   </a>
 </sup>
 
-> Structure and organize the UI
+> Structure and organize the UI using columns
 
 ## Installation
 
@@ -69,6 +69,8 @@ import * as packageJson from '../package.json';
 > Also notice that in Item 1 we are making use of the [UiViewport](./packages-foundation-docs-viewport) to render/remove it depending on the breakpoint
 
 ### 3. Using GridBreakpointsDistribution prop
+
+This prop give you the ability to set a number of cols/rows depending on the breakpoint.
 
 <Playground>
   <UiGrid cols={{ small: 1, medium: 2, large: 3, xlarge: 3 }} rows={2} colsGap={20} rowsGap={10}>
