+++ packages/grid/src/index.ts
@@ -1,2 +1,3 @@
 export * from './grid';
 export * from './grid-item';
+export type { UiGridProps, GridBreakpointsDistribution } from './types';
