+++ packages/grid/src/private/calculate-grid-template.ts
@@ -1,11 +1,4 @@
-import { GridBreakpointsDistribution } from '../';
-
-export const getGridTemplate = (
-  spans: number | GridBreakpointsDistribution = 1,
-  size = '1fr',
-  template: 'cols' | 'rows'
-): string => {
-  return `${template === 'cols' ? 'grid-template-columns' : 'grid-template-rows'}: repeat(${spans}, ${
+export const getGridTemplate = (spans = 1, size = '1fr', template: 'cols' | 'rows'): string =>
+  `${template === 'cols' ? 'grid-template-columns' : 'grid-template-rows'}: repeat(${spans}, ${
     size === '' ? '1fr' : size
   });`;
-};
